multiple fix and upgrade
1. VDialog accepts a list of VDialogButtons
2. fix: Viewport horizontal calculation
3. support to specify init scene when creating stage
4. hpadding and vpadding accepts a double value

+++ src/main/java/io/vproxy/vfx/ui/layout/HPadding.java
@@ -4,7 +4,7 @@ import javafx.geometry.Insets;
 import javafx.scene.layout.Pane;
 
 public class HPadding extends Pane {
-    public HPadding(int padding) {
+    public HPadding(double padding) {
         setVisible(false);
         setWidth(0);
         setPrefWidth(0);
